+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- A canceled client stream blocks further request
+
 ## [v0.3.3] - 2020-12-13
 
 ### Added