+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Date picker for `google.protobuf.Timestamp` fields
 
+### Fixed
+- A canceled client stream blocks further requests
+
 ## [v0.3.3] - 2020-12-13
 
 ### Added