+++ CHANGELOG.md
@@ -6,8 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Request type icons, in-flight indicator and network out/in count
+
 ### Fixed
-- A canceled client stream blocks further request
+- A canceled client stream blocks further requests
 
 ## [v0.3.3] - 2020-12-13